Added basis for idea
This commit includes a lot of stuff that has been made beforehand to test an idea. It has a Random level generator, a player object that can move and collide, a working game engine with a viewport, and also a class that handles sprites with animations. The code will be optimized when we go along with the project.