Player can now attack!
This came with many other changes. First of all the update method in the Game class had to be changed into an iterator loop, to be able to remove stuff when iterating. Also added a temporary list to store objects created during the update method. Finally, every GameObject now has its own ID which makes it easier to
refer to an individual object.